Move the inst_scc.rb content to library
to properly count the code coverage
